Use timelines from the config
